pyproject.toml
[project]
name = "mini-ansible"
version = "0.1.0"
description = "A lightweight Ansible-like tool for provisioning and configuration over SSH."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paramiko>=3.5.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
mini-ansible = "mini-ansible.cli:main"